Rename more non-class modules to snake case
