+++ src/content-scripts/modules/annotation-locator.ts
