+++ src/content-scripts/modules/app-component.tsx
