+++ src/content-scripts/modules/highlight-manager.ts
